change fitness value to float

+++ app/models.py
@@ -43,7 +43,7 @@ class TeamMember(models.Model):
 class FitnessStat(models.Model):
     tracked_on = models.DateField(null=False, blank=False)
     metric = models.CharField(max_length=256, null=False, blank=False, choices=fitness_metrics)
-    value = models.IntegerField(null=False, blank=False)
+    value = models.FloatField(null=False, blank=False)
     team_member = models.ForeignKey(TeamMember, on_delete=models.CASCADE)
     file = models.FileField(null=True, blank=True)
 
@@ -54,11 +54,3 @@ class FitnessStat(models.Model):
     def __str__(self):
         return f'{self.metric} on {self.tracked_on} of {self.team_member}'
 
-# @admin.action(description='Mark selected stories as published')
-# def make_published(modeladmin, request, queryset):
-#     queryset.update(status='p')
-#
-# class ArticleAdmin(admin.ModelAdmin):
-#     list_display = ['title', 'status']
-#     ordering = ['title']
-#     actions = [make_published]
